factor out a parent gate verifier analysis, so it can be reused with merge_rotations, etc.
